% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{execute}
\alias{execute}
\title{Execute the validation study}
\usage{
execute(
  databaseDetails,
 
    restrictPlpDataSettings = PatientLevelPrediction::createRestrictPlpDataSettings(sampleSize
    = NULL),
  validationSettings = PatientLevelPrediction::createValidationSettings(recalibrate =
    NULL),
  outputFolder,
  createCohorts = T,
  runValidation = T,
  packageResults = T,
  minCellCount = 5,
  logSettings = PatientLevelPrediction::createLogSettings(verbosity = "INFO", logName =
    "validatePLP")
)
}
\arguments{
\item{databaseDetails}{Database details for the validation created using \code{PatientLevelPrediction::createDatabaseDetails()}}

\item{restrictPlpDataSettings}{Extras data settings such as sampling created using \code{PatientLevelPrediction::createRestrictPlpDataSettings()}}

\item{validationSettings}{Settings for the validation such as whether to recalibrate}

\item{outputFolder}{Name of local folder to place results; make sure to use forward slashes
(/)}

\item{createCohorts}{Whether to create the cohorts for the study}

\item{runValidation}{Whether to run the valdiation models}

\item{packageResults}{Whether to package the results (after removing sensitive details)}

\item{minCellCount}{The min count for the result to be included in the package results}

\item{logSettings}{Settings for the logging created using \code{PatientLevelPrediction::createLogSettings()}}
}
\description{
Execute the validation study
}
\details{
This function will execute the sepcified parts of the study
}
