% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PackageResults.R
\name{packageResults}
\alias{packageResults}
\title{Package the results for sharing with OHDSI researchers}
\usage{
packageResults(outputFolder, databaseName, minCellCount = 5)
}
\arguments{
\item{outputFolder}{Name of folder containing the study analysis results}

\item{databaseName}{Name of the database being validated
(/)}

\item{minCellCount}{The minimum number of subjects contributing to a count before it can be included in the results.}
}
\description{
Package the results for sharing with OHDSI researchers
}
\details{
This function packages the results.
}
