% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extraPredictFunctions.R
\name{predictNonPlpGlm}
\alias{predictNonPlpGlm}
\title{Prediction for existing GLM}
\usage{
predictNonPlpGlm(plpModel, data, cohort)
}
\arguments{
\item{plpModel}{The model being applied}

\item{data}{The new data}

\item{cohort}{The new population}
}
\value{
The population with an extra column 'value' corresponding to the patients risk
}
\description{
Prediction for existing GLM
}
\details{
This applies the existing models and calcualtes the risk for a population
}
