% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SkeletonPredictionValidationStudy.R
\docType{package}
\name{SkeletonPredictionValidationStudy}
\alias{SkeletonPredictionValidationStudy}
\title{SkeletonPredictionValidationStudy}
\description{
A package for validating prediction models using data in the OMOP CDM
}
